+++ NEWS.md
@@ -5,6 +5,14 @@ the EndBASIC language definition nor the API exposed by this crate.  Expect
 them to change at any time (especially the Rust API).  Version numbers will
 not adhere to semantic versioning until 1.0.0.**
 
+## Changes in version X.Y.Z
+
+**STILL UNDER DEVELOPMENT; NOT RELEASED YET.**
+
+*   Refactored the `HELP` command to only print a summary of topics by
+    default.  As part of this, topics can now be looked up using prefixes of
+    their names.
+
 ## Changes in version 0.6.0
 
 **Released on 2021-02-19.**